accepting floating point

+++ usefulprograms/ratio.cpp
@@ -3,11 +3,11 @@
 using namespace std;
 
 int main(int argc, char *argv[]){
-	int num;
-	int percent;
+	float num;
+	float percent;
 	cout << "Enter number: ";
 	cin >> num;
 	cout << "Enter percent: ";
 	cin >> percent;
-	cout << num*(((double)percent)/100) << endl;
+	cout << num*(percent/100) << endl;
 }
